+++ packages/common/src/models/index.ts
@@ -5,9 +5,5 @@ export * from "./class";
 export * from "./term";
 export * from "./course";
 export * from "./section";
-export * from "./sectionNew";
-export * from "./courseNew";
-export * from "./classNew";
 export * from "./grade-distribution";
 export * from "./enrollment-history";
-export * from "./termNew";
